move static_assets handler and other utils
