+++ crates/rari_core/src/routing/mod.rs
@@ -1,2 +1,3 @@
 //! Routing primitives shared across backends (segment types, param values).
+pub mod api_error;
 pub mod types;
