+++ crates/rari_core/src/utils/mod.rs
@@ -1,3 +1,4 @@
 pub mod cast;
 pub mod float;
 pub mod path;
+pub mod path_validation;
